Substantially revised the ASN.1 in sections 4-6
The previous draft re-defined existing X.509 certificate fields.  This
commit defines new public key and signature formats only, so that the
original field definitions can be used.  Other cosmetic changes were
made as well.  Sections 7 and later still need to be updated to reflect
the changes made in this commit.