Only define a single OID for composite objects
Separate OIDs are only necessary for keys and signature algorithms
when the signature algorithm can be combined with a pre-hash (e.g.
ecdsa-with-sha256).  This doesn't apply to composite signatures
since the pre-hash will be specific to each of the component
algorithms.

This change also uses the SIGNATURE-ALGORITHM and PUBLIC-KEY object
types of RFC 5912.

Closes #50